// apps/api/src/groups/dto/update-group.dto.ts
import { ValidationSchema } from '@douglasneuroinformatics/libnest';
import { $UpdateGroupData } from '@opendatacapture/schemas/group';
import type { GroupSettings, GroupType, UpdateGroupData } from '@opendatacapture/schemas/group';

@ValidationSchema($UpdateGroupData)
export class UpdateGroupDto implements UpdateGroupData {
  accessibleInstrumentIds?: string[];
  hospitals?: string[];
  name?: string;
  settings?: Partial<GroupSettings>;
  type?: GroupType;
}